// model/updateOrderTaxRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* create new taxes for an existing order
*/
export class UpdateOrderTaxRequest {
    /**
    * The amount to be collected for tax in cents
    */
    'amount'?: number;
    /**
    * description or tax\'s name
    */
    'description'?: string;
    'metadata'?: { [key: string]: object; };

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: object; }"
        }    ];

    static getAttributeTypeMap() {
        return UpdateOrderTaxRequest.attributeTypeMap;
    }
}