// model/update-order-tax-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * create new taxes for an existing order
 * @export
 * @interface UpdateOrderTaxRequest
 */
export interface UpdateOrderTaxRequest {
    /**
     * The amount to be collected for tax in cents
     * @type {number}
     * @memberof UpdateOrderTaxRequest
     */
    'amount'?: number;
    /**
     * description or tax\'s name
     * @type {string}
     * @memberof UpdateOrderTaxRequest
     */
    'description'?: string;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof UpdateOrderTaxRequest
     */
    'metadata'?: { [key: string]: object; };
}